Use a higher CPU usage method on WIN32 to get better timer precision
When a timer needs to run, floor the sleep time instead of using the
ceiling. This means on waking up the timer may not yet need to run
but that the timer wont be slept through.

This commit is a test to check on how high the resulting CPU usage is,
to check if it's acceptable or not.