Sculpt: Fix #104022: Expand Mask inverted fill
A small new feature that clears the mask completely on
the current mesh island.

When hovering outside the object the Expand operator
is running on, it will get floodfilled with the mask/face set.
Previously, when the Invert option was used, it wouldn't also
clear the mask/face set when hovering outside the object.