Fix #111613: Gesture selection does not deselect when elements hidden
When objects/collections are hidden, `BASE_SELECTED` flag is cleared in
`BKE_base_eval_flags`. When deselecting in viewport, this flag is
considered, otherwise selection is skipped. If nothing is changed during
select/box/lasso select, notifiers are not sent to other regions. Due to
above reasons, selection of hidden object is skipped.
To fix this, send notifiers to outliner whenever operation is executed.

Would it affect the performace? I'm not sure. In that case we can think
of alternative solution.