Fix: USD import: deform group name collision.
This addresses a bug when importing USD skeleton data, where
deform groups are misnamed due to name collisions with existing
mesh attributes. The incorrectly named deform groups fail to
bind to the armature animation.

The issue is that the call to BKE_object_defgroup_add_name()
assigns a unique name to the deform group if an attribute
with the same name already exists in the imported mesh.

The solution implemented in this commit is to remove attributes
with colliding names before creating the deform groups.

To reproduce, import the attached arm_test.usda, which contains
skeletal joints Hand, Shoulder and Elbow as well as mesh
attributes with the same names.  Note that the imported deform
groups are renamed Hand.001, Shoulder.001 and Elbow.001 and that
the mesh fails to bind to the armature animation.

This issue would be prone to happening on round trips
Blender -> USD -> Blender, because the Blender deform groups would
be exported as both mesh attributes and USD skeleton binding data.