Shading Nodes: Refactor Add menu
Refactor the Add menu in Shading nodes, with manually created menus,
inspired by Geometry Nodes and more recently the Compositor.

Minor sorting adjustments by splitting categories in groups, with
separators in between groups, and sorted alphabetically. Shading
node group assets are also populated in the menus.

This is the first part of the re-organization of the Add menu project,
proposal coming in another patch after this.

This patch also:
* Move the Layout menu to a more generic bit that can be shared
  with Geometry Nodes and the Compositor.
* Add a poll to add_node_type so nodes show up or not by mode
  or render engine.

Part of #111746