Cycles: remove redundant implementation of threads
Currently, the pthread implementation of threads in cycles is used only for MacOS,
and for all other operating systems, the std::thread implementation is used.

This pthread implementation is needed for MacOS in order to increase the
stack size for threads, because MacOS's default thread stack size is too small
for embree. There is however no good way of doing this with C++'s std::thread
since the C++ specification does not define a stack for threads

Therefore, remove std::thread alternative implementation from cycles and just
use just one implementation for all systems, and have consistent stack sizes.

This also makes the binary slightly smaller.