USD: skip exporting deform group attributes.
No longer exporting mesh attributes that correspond to deform
groups.

This is to avoid potential name collisions when importing the
USD back into Blender, as described in

https://projects.blender.org/blender/blender/pulls/111747

This preemptively addresses an issue which will arise when
exporting armatures to USD is supported, because the deform group
attributes and the UsdSkel mesh skinning data will be redundant.

In the future we might consider refining this behavior with a USD
export option to allow exporting some vertex group attributes but
not others.