Code quality: Add type-safe C++ functions to cast ID types
`reinterpret_cast()`s are annoying to type, contain redundant
information (e.g. `const` or pointer/ref) and remove type safety.
Non-obvious issue: Being annoying to type leads people to avoid the
casting and use `&some_id->id` instead, which can introduce null-pointer
dereferences when `some_id` is null (null may still be a valid value for
the result of the expression). So making casts easy to type can in fact
prevent bugs.

This adds simple ID-type casting functions from IDs to the "derived"
type and vise versa. Main benefits:
- Contain static type checks to ensure the requested derived type can
  actually be used as ID type (e.g. `id_cast<Sequence>(...)` or
  `id_cast<Bone>(...) won't compile, unlike `reinterpret_cast`).
- Preserve `const` correctness implicitly.
- Works for both pointers and references implicitly.
- Easy to search for and change unlike all possible other casts (useful
  in case we switch to C++ style inheritance, for example).
- Easy to write.

To cast an ID pointer/reference to a derived type, the derived type
needs to be passed as template parameter. To cast to an ID
pointer/reference, no template parameter is required.

With this we might want to make it a good-practice or even guideline to
use the cast over alternatives.