Cycles: Improve compilation error logs for Metal kernels
The Metal backend combines all kernel sources into a single string
and passes it to the compiler. Doing so natively has a disadvantage
of making it hard to find sources of mistakes in code when working
on the kernel source as the source file name and line number info
is lost. For example, the error would look like:

  program_source:187004:3: error: use of undeclared identifier 'metalrt_intersection_point_shadow_force_error'
    metalrt_intersection_point_shadow_force_error(launch_params_metal,
    ^

This patch makes it so the #line pragmas are inserted into the
combined source source code, so that the compiler errors are easier
to find immediately. For example the error from above becomes

  source/kernel/device/metal/kernel.metal:809:3: error: use of undeclared identifier 'metalrt_intersection_point_shadow_force_error'
    metalrt_intersection_point_shadow_force_error(launch_params_metal,
    ^

The code is a slightly modified functionality from the source
processing used for OpenCL in Blender 2.80.