Fix #112014: Crash making node group with multiple input links.
e071288ab2 changed the "make node group" operator for new interfaces,
and in the process added incorrect lines removing a link when a group
input exists already. This crashes when there are multiple sockets
in selected nodes linked to the same non-selected output.

The link should not be removed, this is just lazy-initialization of the
group sockets.