Improve punctual shadow cubeface transition
This enlarges the projection frustum to make sure all possible
rays for a receiver pixel can be traced inside one cube face.

This removes a great deal of complexity from the tracing and
reduces the artifacts.

This also add an offset in the cube for area light so that
nearby occluders do not get clipped.