WIP: UI: Viewport: Show Add menu if nothing is selected
In Object and Mesh modes, show the Add menu (and Paste in Object mode).

Similar to how it's done in Node Editor.

The code for it was already there, but commented out because it was
seen as not useful. However, none of the menu entries work without
objects selected (except Paste), so the menu is already not useful as is.