USD: Use pre-modified mesh for skinned meshes.
Now using the pre-modified mesh to export the mesh
in its rest pose.  This replaces the use of the
ModifierDisabler class, which was causing some
animation data to be dropped.  Using the modifier
disabler as an approach could be revisited in the
future.  Also, moved the mesh creation function for
blend shapes has been moved to a new
get_shape_key_basis_mesh() function.