Fix: Cycles Principled BSDF specular component wrong IOR when backfacing
is visible when Transmission value is between 0 and 1.
The behaviour is actually not well defined for specular layer when
coming from inside, but this is at least consistent with the
transmission component and also consistent with the current behaviour in
OSL. Could probably save one closure allocation in svm later.