I18n: revert context of items in the Add Modifier menu to default
The Add Modifier menu used to be an operator which called a menu on
invocation (`object.modifier_add`, called through
`operator_menu_enum()`). Each item used the `default` translation
context.

This has changed to using an actual menu, in which each possible type
of modifier is displayed using `layout.operator()`. This results in
using the `default_operator` translation context.

Since the context has changed, the translations must be updated, but
this poses two problems:
1. The old way of calling the menu still exists if one calls the
operator manually, so the translation is duplicated for nothing.
2. Some of the messages already exist with the Operator context, and
mean something else (e. g. "Bevel" is the imperative form of the verb,
instead of the noun).

This commit fixes those issues by restoring the `default` context in
the interface of the new menu.