UI/I18N: Fix a couple of issues in new 'Add Modifiers' menu.
The recent change to the 'Add Modifier' created at least two issues:
* A complete duplicate of UI info for each modifier (its name and icon),
  now existing in both the RNA enum definition of modifiers types, and
  in the pyhton UI code.
* An implicit duplication of these UI names in two different translation
  contexts, since the ones from the enum use the default one, while
  explicit labels passed to the `layout.operator` UI API get assigned a
  default 'operator' context. See PR !112246 for details about this.

Both issues can be easily solved by making the new python code for these
menus a bit smarter. Adding a helper function that adds the `add_modifier`
operator with the right parameters, just based on the operator type.

Both names (labels) and icons can be found in the enum property
`bl_rna` definition itself then.

This change:
* Avoids duplicating UI info.
* Fixes translation context mismatch, by forcing the usage of the
  default one also from the python code.
* Makes code less verbose and overall more readable.

NOTE: An attempt has been made to use the `get_name` callback of
operator types to automatically return the right name based on the
defined type, but this is currently utterly failing with regular
layout-based UI code. This will be reported and handled separately.