Mesh: add test for isomorphic meshes
In many cases when updating some mesh algorithm, the resulting mesh
might have different indices. Current tests could not detect this and
would complain that the two meshes are different. The meshes_isomorphic
function attempts to solve this problem, by finding a re-ordering of
indices such that the two meshes become equal.