Allow setting weights (Ctrl+X) without paintmask enabled
Compared to setting vertex colors with Ctrl+X (which does not have the
requirement of having vertex/face mask enabled) this was inconsistent.
It is also not needed to check if paintmask is enabled, the code checks
for this just fine.

So now remove the overly-strict `mask_paint_poll` poll (and replace with
more appropriate `weight_paint_mode_poll`).

NOTE: `mask_paint_poll` is now unused, could be removed