deps_builder: facilitate cve-bin-tool in a CI environment
There were a few issues

1 - The command line was shown for running cve-bin-tool leaking the
NVD key which could end up in the CI logs, which is undesirable,
this is fixed by setting the COMMENT for a generic message.

2 - it's slow to run, and ninja buffers all output making it a bit
difficult to see what was happening, this is fixed by adding
USES_TERMINAL which puts it in the console pool for ninja so you
get output as it happens, it does nothing for all other generators

3 - All deps were downloaded and msys2 was setup, none of that is
needed if you just want to run the CVE check or HTML version report
for this a new option is added PACKAGE_DOWNLOAD_SOURCES which is
ON by default, but can be turned off to facilitate the reports.

4 - My pipeline needed stable filenames and the json report, so i
added an additional target for that.

---

My pipeline is currently outputting to #blender-status-cve (but won't
run on a daily schedule until this pull lands)