Assets: Replace most asset handle usages in BPY with asset representation
The asset handle type is supposed to be replaced by the asset
representation type. It is designed for the asset system as opposed to
the file browser backend. This commit adds members to its RNA type to
query data, replaces most usages of asset handle with asset
representation, and removes now unnecessary BPY/RNA properties and
functions.

Part of #102877 and #108806.