Creator: Remove Disabling SSBO Command Line Argument
Blender has the option to disable SSBO support. This was accessible
via the command line option `--debug-gpu-disable-ssbo`.

For Blender 4.0 SSBO is a hard requirement as it is needed for Workbench
and SSBO support is part the new OpenGL requirement (4.3 core).
This PR removes the command line option as it makes no sense to
have it anymore.