Add extra subtypes for float array
I think these are the most useful ones that are missing.
I left out matrix because it doesn't play nice with UI so I figure nobody making custom properties would need it.
I also left out coordinates because as far as I can tell it's the same thing as XYZ.