Fix #112580: Limit Constraint with 'Affect Transform' not working properly in 'World Space'
The problem is observed with the Limit Distance and Limit Location
constraints.

There is an incorrect usage of td->mtx and td->smtx when converting
TransData space from local to global.

In this case, the code is concatenating matrices instead of converting
the location component space.

Since the Limit Distance and Limit Location constraints only require
the location component, it is not necessary to convert the rotation and
scale components.

So, the solution is to convert the location component space instead of
concatenating matrices.