Fix #104554: 2d paint does not paint with linear color on 32bit images
When painting in the Image Editor on float images, the brush color was
not converted from sRGB to linear, leading to differences painting with
the same brush in the 3DView vs. the Image Editor.

Now corrected by piping the brush color through
srgb_to_linearrgb_v3_v3() when painting on 32bit images.

Should probably go into LTS as well.