Asset shelf: Better region snapping behavior, "Show Names" changes size
Rather than always snapping the region size to the closest multiple of
the row height, remember the amount of rows displayed after the user
resized the region, and try to preserve that. This gives a lot more
predictable behavior, especially when the "Show Names" option is toggled
on and off, and the region resizes in response.

Also contains a fix for the tile height calculation for when a label is
shown.