Fix 110976: Sequencer Scope Artifacts on NVIDIA 30/40 Series
This PR fixes an issue with the NVIDIA 3000/4000 series cards. The cause
is that when blending is turned off and the fragment output has an alpha
of 0.0 it seems to internally optimize the fragment away, leading. This
is a guess of course.

By setting the Alpha to 1.0 for scopes and enabling alpha blending when
drawing the meta data text seems to fix the issue.

As I am not aware of all the different user setting please review this
carefully.