UI: tweaks to Library Override and Relations in Object menu
Remove duplicate `object.make_override_library` entry in Object menu.

It was in both the "Relations" and the "Library Override" menu.
Remove it from "Relations".

Also add an icon to Library Override (similar to Assets), add
separator and move it up since it is now the intended way of
handling relations.